feat(newApi): wip -> facade
